Renamed and reorganized some of the classes.
