+++ include/scran/quality_control/ComputeMedianMad.hpp
@@ -1,5 +1,5 @@
-#ifndef SCRAN_COMPUTE_MED_MAD_H
-#define SCRAN_COMPUTE_MED_MAD_H
+#ifndef SCRAN_COMPUTE_MEDIAN_MAD_H
+#define SCRAN_COMPUTE_MEDIAN_MAD_H
 
 #include "../utils/macros.hpp"
 
@@ -10,10 +10,9 @@
 #include <cstdint>
 
 #include "utils.hpp"
-#include "../utils/block_indices.hpp"
 
 /**
- * @file ComputeMedMad.hpp
+ * @file ComputeMedianMad.hpp
  *
  * @brief Compute the median and MAD from an array of values.
  */
@@ -27,7 +26,7 @@ namespace scran {
  * and uses the median to then compute the median absolute deviation (MAD) from that array.
  * We support calculation of medians and MADs for separate "blocks" of observations in the same array.
  */
-class ComputeMedMad {
+class ComputeMedianMad {
 public:
     /**
      * Whether to compute the median and MAD after log-transformation of the values.
@@ -76,13 +75,48 @@ public:
 
 public:
     /**
+     * @tparam Block Integer type, containing the block IDs.
+     * 
+     * @param n Number of observations.
+     * @param[in] block Pointer to an array of block identifiers.
+     * If provided, the array should be of length equal to `n`.
+     * Values should be integer IDs in \f$[0, N)\f$ where \f$N\f$ is the number of blocks.
+     * If a null pointer is supplied, all observations are assumed to belong to the same block.
+     * 
+     * @return Vector of indices specifying the start position for each block on the workspace buffer,
+     * if observations from successive blocks were arranged contiguously onto the buffer.
+     */
+    template<typename Block>
+    static std::vector<int> compute_block_starts(size_t n, const Block* block) {
+        std::vector<int> starts;
+
+        for (size_t i = 0; i < n; ++i) {
+            size_t candidate = block[i] + 1;
+            if (candidate > starts.size()) {
+                starts.resize(candidate);
+            }
+            ++starts[block[i]];
+        }
+        
+        int sofar = 0;
+        for (auto& s : starts) {
+            int last = sofar;
+            sofar += s;
+            s = last;
+        }
+
+        return starts;
+    }
+
+public:
+    /**
      * @tparam Input Numeric type for the input.
      * @tparam Buffer Floating-point type for the buffer.
      *
      * @param n Number of observations.
      * @param[in] metrics Pointer to an array of observations of length `n`.
      * NaNs are ignored.
-     * @param buffer Pointer to an array of length `n`.
+     * @param buffer Pointer to an array of length `n`, to be used as a workspace.
      *
      * @return Median and MAD for the `metrics`.
      */
@@ -104,37 +138,69 @@ public:
     }
 
     /**
+     * @tparam Block Integer type, containing the block IDs.
+     * @tparam Input Type of observation, numeric.
+     * @tparam Buffer Floating-point type for the buffer.
+     *
+     * @param n Number of observations.
+     * @param[in] block Pointer to an array of block identifiers.
+     * If provided, the array should be of length equal to `n`.
+     * Values should be integer IDs in \f$[0, N)\f$ where \f$N\f$ is the number of blocks.
+     * If a null pointer is supplied, all observations are assumed to belong to the same block.
+     * @param[in] metrics Pointer to an array of observations of length `n`.
+     * @param buffer Pointer to an array of length `n`, to be used as a workspace.
+     *
+     * @return Median and the MAD for each block in `metrics`.
+     */
+    template<typename Input, typename Block, typename Buffer>
+    Results run_blocked(size_t n, const Block* block, const Input* metrics, Buffer* buffer) const {
+        if (block) {
+            return run_blocked(n, block, compute_block_starts(n, block), metrics, buffer);
+        } else {
+            return run(n, metrics, buffer);
+        }
+    }
+
+    /**
+     * @tparam Block Integer type, containing the block IDs.
      * @tparam Input Numeric type for the input.
      * @tparam Buffer Floating-point type for the buffer.
      *
      * @param n Number of observations.
-     * @param by_block Vector of length equal to the number of blocks.
-     * Each entry is a vector of integers, containing the indices of the elements in `metrics` in the corresponding block.
-     * All indices should lie in `[0, n)`.
+     * @param[in] block Optional pointer to an array of block identifiers.
+     * If provided, the array should be of length equal to `ncells`.
+     * Values should be integer IDs in \f$[0, N)\f$ where \f$N\f$ is the number of blocks.
+     * If a null pointer is supplied, all observations are assumed to belong to the same block.
+     * @param starts Vector of start positions for each block,
+     * typically generated by calling `compute_block_starts()` on `n` and `block`.
      * @param[in] metrics Pointer to an array of observations of length `n`.
-     * NaNs are ignored.
-     * @param buffer Pointer to an array of length `n`.
+     * @param buffer Pointer to an array of length `n`, to be used as a workspace.
      *
      * @return Medians and MADs for each block in `metrics`.
      */
-    template<typename Input, typename Buffer>
-    Results run_blocked(size_t n, const std::vector<std::vector<size_t> >& by_block, const Input* metrics, Buffer* buffer) const {
-        Results output(by_block.size());
-        output.log = log;
-
-        for (size_t g = 0; g < by_block.size(); ++g) {
-            auto copy = buffer;
-            const auto& curblock = by_block[g];
-            for (auto s : curblock) {
-                if (!std::isnan(metrics[s])) {
-                    *copy = metrics[s];
-                    ++copy;
-                }
+    template<typename Input, typename Block, typename Buffer>
+    Results run_blocked(size_t n, const Block* block, std::vector<int> starts, const Input* metrics, Buffer* buffer) const {
+        if (block) {
+            // Unscrambling into the buffer.
+            auto ends = starts;
+            for (size_t i = 0; i < n; ++i) {
+                auto& pos = ends[block[i]];
+                buffer[pos] = metrics[i];
+                ++pos;
             }
-            compute(copy - buffer, buffer, output.medians[g], output.mads[g]);
-        }
 
-        return output;
+            // Using the ranges on the buffer.
+            size_t nblocks = starts.size();
+            Results output(nblocks);
+            output.log = log;
+            for (size_t g = 0; g < nblocks; ++g) {
+                compute(ends[g] - starts[g], buffer + starts[g], output.medians[g], output.mads[g]);
+            }
+
+            return output;
+        } else {
+            return run(n, metrics, buffer);
+        }
     }
 
 public:
@@ -167,14 +233,9 @@ public:
      * @return Median and the MAD for each block in `metrics`.
      */
     template<typename Block, typename Input>
-    Results run_blocked(size_t n, const Block * block, const Input* metrics) const {
-        if (block) {
-            BlockIndices by_block = block_indices(n, block);
-            std::vector<double> buffer(n);
-            return run_blocked(n, by_block, metrics, buffer.data());
-        } else {
-            return run(n, metrics);
-        }
+    Results run_blocked(size_t n, const Block* block, const Input* metrics) const {
+        std::vector<double> buffer(n);
+        return run_blocked(n, block, metrics, buffer.data());
     }
 
 private:
@@ -247,84 +308,6 @@ private:
 
         return medtmp;
     }
-
-public:
-    /**
-     * @brief Define outlier filters from the median and MAD.
-     *
-     * Given an array of values, outliers are defined as those that are more than some number of median absolute deviations (MADs) from the median value.
-     * By default, we require 3 MADs, which is motivated by the low probability (less than 1%) of obtaining such a value under the normal distribution.
-     * Outliers can be defined in both or either direction.
-     */
-    struct FilterOutliers {
-    public:
-        /**
-         * Should low values be considered as potential outliers?
-         * If `false`, no lower threshold is applied when defining outliers.
-         */
-        bool lower = true;
-
-        /**
-         * Should high values be considered as potential outliers?
-         * If `false`, no upper threshold is applied when defining outliers.
-         */
-        bool upper = true;
-
-        /**
-         * Number of MADs to use to define outliers.
-         */
-        int num_mads = 3;
-
-        /**
-         * Minimum difference from the median to define outliers.
-         * If the median and MADs are log-transformed, this difference is interpreted as a unit on the log-scale.
-         */
-        double min_diff = 0;
-
-    private:
-        double sanitize(double proposed, bool log) const {
-            if (log) {
-                if (std::isinf(proposed)) {
-                    if (proposed < 0) {
-                        proposed = 0;
-                    }
-                } else {
-                    proposed = std::exp(proposed);
-                }
-            } 
-            return proposed;
-        }
-
-    public:
-        /**
-         * @param x Median and MAD computed from `ComputeMedMad::run()` or `ComputeMedMad::run_blocked()`.
-         * @return Outlier filter thresholds defined in terms of MADs from the median.
-         */
-        FilterThresholds run(ComputeMedMad::Results x) const {
-            size_t nblocks = x.medians.size();
-            for (size_t b = 0; b < nblocks; ++b) {
-                auto med = x.medians[b];
-                auto mad = x.mads[b];
-
-                double& lthresh = x.medians[b];
-                double& uthresh = x.mads[b];
-
-                if (std::isnan(med)) {
-                    lthresh = std::numeric_limits<double>::quiet_NaN();
-                    uthresh = std::numeric_limits<double>::quiet_NaN();
-                } else {
-                    auto delta = std::max(min_diff, num_mads * mad);
-                    lthresh = (lower ? sanitize(med - delta, x.log) : -std::numeric_limits<double>::infinity());
-                    uthresh = (upper ? sanitize(med + delta, x.log) : std::numeric_limits<double>::infinity());
-                }
-            }
-
-            FilterThresholds output;
-            output.lower = std::move(x.medians);
-            output.upper = std::move(x.mads);
-            return output;
-        }
-    };
 };
 
 }
