+++ tests/src/quality_control/ComputeMedianMad.cpp
@@ -1,9 +1,7 @@
 #include <gtest/gtest.h>
 #include "../utils/macros.h"
 
-#include "../data/data.h"
-
-#include "scran/quality_control/ComputeMedMad.hpp"
+#include "scran/quality_control/ComputeMedianMad.hpp"
 
 std::vector<double> even_values = { 
     0.761164335, 0.347582428, 0.430822695, 0.888530395, 0.627701241, 0.678912751,
@@ -12,8 +10,8 @@ std::vector<double> even_values = {
     0.554883985, 0.009624974, 0.215816610
 };
 
-TEST(ComputeMedMad, BasicTests) {
-    scran::ComputeMedMad is;
+TEST(ComputeMedianMad, BasicTests) {
+    scran::ComputeMedianMad is;
     auto stats = is.run(even_values.size(), even_values.data());
 
     EXPECT_FALSE(stats.log);
@@ -33,8 +31,8 @@ TEST(ComputeMedMad, BasicTests) {
     EXPECT_FLOAT_EQ(stats2.mads[0], 0.2731710715);
 }
 
-TEST(ComputeMedMad, EdgeTests) {
-    scran::ComputeMedMad is;
+TEST(ComputeMedianMad, EdgeTests) {
+    scran::ComputeMedianMad is;
 
     std::vector<double> empty;
     auto stats = is.run(empty.size(), empty.data());
@@ -70,8 +68,8 @@ TEST(ComputeMedMad, EdgeTests) {
     }
 }
 
-TEST(ComputeMedMad, LogTests) {
-    scran::ComputeMedMad is;
+TEST(ComputeMedianMad, LogTests) {
+    scran::ComputeMedianMad is;
     is.log = true;
 
     auto isres = is.run(even_values.size(), even_values.data());
@@ -102,7 +100,7 @@ TEST(ComputeMedMad, LogTests) {
     }
 }
 
-TEST(ComputeMedMad, BlockTests) {
+TEST(ComputeMedianMad, BlockTests) {
     std::vector<int> block = {
         0, 1, 2, 3, 
         1, 0, 2, 3,
@@ -112,7 +110,7 @@ TEST(ComputeMedMad, BlockTests) {
         1
     };
 
-    scran::ComputeMedMad is;
+    scran::ComputeMedianMad is;
     auto isres = is.run_blocked(even_values.size(), block.data(), even_values.data());
 
     EXPECT_EQ(isres.medians.size(), 4);
@@ -131,105 +129,16 @@ TEST(ComputeMedMad, BlockTests) {
         EXPECT_EQ(isres.mads[i], is2res.mads[0]);
     }
 
-    // NULL blocked is no-op.
+    // NULL blocked falls back to single-batch processing.
     {
         auto isres_none = is.run_blocked(even_values.size(), static_cast<int*>(NULL), even_values.data());
         auto ref = is.run(even_values.size(), even_values.data());
         EXPECT_EQ(ref.medians, isres_none.medians);
         EXPECT_EQ(ref.mads, isres_none.mads);
+
+        std::vector<double> buffer(even_values.size());
+        auto isres_none2 = is.run_blocked(even_values.size(), static_cast<int*>(NULL), {}, even_values.data(), buffer.data());
+        EXPECT_EQ(ref.medians, isres_none2.medians);
+        EXPECT_EQ(ref.mads, isres_none2.mads);
     }
-}
-
-TEST(ComputeMedMad, OutlierFilterSimple) {
-    scran::ComputeMedMad is;
-    auto stat = is.run(even_values.size(), even_values.data());
-
-    // Manual check.
-    {
-        scran::ComputeMedMad::FilterOutliers filt;
-        auto thresholds = filt.run(stat);
-        EXPECT_DOUBLE_EQ(thresholds.lower[0], stat.medians[0] - stat.mads[0] * 3);
-        EXPECT_DOUBLE_EQ(thresholds.upper[0], stat.medians[0] + stat.mads[0] * 3);
-    }
-
-    // Turns off on request.
-    {
-        scran::ComputeMedMad::FilterOutliers filt;
-        filt.lower = false;
-        filt.upper = false;
-        auto thresholds = filt.run(stat);
-
-        EXPECT_TRUE(std::isinf(thresholds.lower[0]));
-        EXPECT_TRUE(thresholds.lower[0] < 0);
-        EXPECT_TRUE(std::isinf(thresholds.upper[0]));
-        EXPECT_TRUE(thresholds.upper[0] > 0);
-    }
-
-    // Respects the minimum difference.
-    {
-        scran::ComputeMedMad::FilterOutliers filt;
-        filt.min_diff = 100;
-        auto thresholds = filt.run(stat);
-        EXPECT_DOUBLE_EQ(thresholds.lower[0], stat.medians[0] - 100);
-        EXPECT_DOUBLE_EQ(thresholds.upper[0], stat.medians[0] + 100);
-    }
-
-    // Handles blocks correctly.
-    {
-        scran::ComputeMedMad::Results stats;
-        stats.medians.push_back(1);
-        stats.mads.push_back(0.2);
-        stats.medians.push_back(2);
-        stats.mads.push_back(0.3);
-
-        scran::ComputeMedMad::FilterOutliers filt;
-        auto thresholds = filt.run(stats);
-        EXPECT_DOUBLE_EQ(thresholds.lower[0], 0.4);
-        EXPECT_DOUBLE_EQ(thresholds.upper[0], 1.6);
-        EXPECT_DOUBLE_EQ(thresholds.lower[1], 1.1);
-        EXPECT_DOUBLE_EQ(thresholds.upper[1], 2.9);
-    }
-}
-
-TEST(ComputeMedMad, OutlierFilterLogged) {
-    scran::ComputeMedMad is;
-    is.log = true;
-    auto stat = is.run(even_values.size(), even_values.data());
-
-    // Manual check.
-    {
-        scran::ComputeMedMad::FilterOutliers filt;
-        auto thresholds = filt.run(stat);
-        EXPECT_DOUBLE_EQ(thresholds.lower[0], std::exp(stat.medians[0] - stat.mads[0] * 3));
-        EXPECT_DOUBLE_EQ(thresholds.upper[0], std::exp(stat.medians[0] + stat.mads[0] * 3));
-    }
-
-    // Mostly zeros.
-    {
-        auto copy = stat;
-        copy.medians[0] = -std::numeric_limits<double>::infinity();
-        copy.mads[0] = 0;
-
-        scran::ComputeMedMad::FilterOutliers filt;
-        auto thresholds = filt.run(copy);
-        EXPECT_EQ(thresholds.lower[0], 0);
-        EXPECT_EQ(thresholds.upper[0], 0);
-    }
-}
-
-TEST(ComputeMedMad, OutlierFilterEdgeCases) {
-    scran::ComputeMedMad::Results stats;
-    stats.medians.push_back(std::numeric_limits<double>::quiet_NaN());
-    stats.mads.push_back(std::numeric_limits<double>::quiet_NaN());
-    stats.medians.push_back(std::numeric_limits<double>::infinity());
-    stats.mads.push_back(0);
-
-    scran::ComputeMedMad::FilterOutliers filt;
-    auto thresholds = filt.run(stats);
-    EXPECT_TRUE(std::isnan(thresholds.lower[0]));
-    EXPECT_TRUE(std::isnan(thresholds.upper[0]));
-    EXPECT_TRUE(std::isinf(thresholds.lower[1]));
-    EXPECT_TRUE(thresholds.lower[1] > 0);
-    EXPECT_TRUE(std::isinf(thresholds.upper[1]));
-    EXPECT_TRUE(thresholds.upper[1] > 0);
 }
